04_agent/strands-python/backend/pyproject.toml
[project]
name = "agent-strands-python"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
  "strands-agents>=0.1.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
  "redis>=5.2.1",
  "pydantic>=2.11.1",
  "anthropic>=0.86.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.28.0",
  "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
